requirements.txt
APScheduler==3.6.3
apturl==0.5.2
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==1.1.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
mypy==0.761
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot==13.4.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smdv==0.1.1
soupsieve==1.9.5
systemd-python==234
Telethon==1.21.1
tornado==6.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
xkit==0.0.0
